+++ example/lib/ui/tasks_tab.dart
@@ -34,7 +34,6 @@ class TasksTab extends StatelessWidget {
           ],
         ),
         body: SolidBuilder<TasksViewModel, TasksState>(
-
           builder: (context, state) {
             final vm = context.solid<TasksViewModel>();
             if (state.tasks.isEmpty && !state.isLoading) {